Support for Nested Stores
